Converting back to float and int.

+++ app/controllers/InitCtrl.php
@@ -44,8 +44,8 @@ class InitCtrl extends \BaseController {
         
        
         ## App versions
-        $return['ios_min_version'] = DB::select('select * from `settings` where `key` = ?', array('ios_min_version'))[0]->value;
-        $return['android_min_version'] = DB::select('select * from `settings` where `key` = ?', array('android_min_version'))[0]->value;
+        $return['ios_min_version'] = (float) DB::select('select * from `settings` where `key` = ?', array('ios_min_version'))[0]->value;
+        $return['android_min_version'] = (int) DB::select('select * from `settings` where `key` = ?', array('android_min_version'))[0]->value;
         
         
         ## Settings
